Serialise CLI output as pretty-printed JSON using Serde
Modify the codegen in 'build.rs' to add Serde attributes to the response
protos of the autogenerated 'walletrpc' module, so that they can be
serialised as JSON by the 'musig-cli' binary, in place of the current
debug-formatted output. This brings its behaviour a bit closer to that
of LND's 'lncli' client (which similarly talks to the daemon via gRPC).
(The stderr output in case of RPC errors remains unaffected.)

For this purpose, add 'serde' & 'serde_with' dependencies to Cargo.toml,
with the former already a transitive dependency, and the latter used to
more easily handle custom serialisation of byte array and enum fields,
via its '#[serde_as(as = ...)]' attribute macro and associated helpers.
This is so that byte arrays such as raw txs, tx IDs and block hashes are
serialised as hex strings (with the byte order of double SHA-256 hashes
such as tx IDs needing to be reversed when displayed as hex). Custom
serialisation is also needed to display enum fields as strings, as they
are raw i32 ordinals in the Prost-generated code (for schema extension).

Also update the expected CLI outputs of the tests.