Fix quadratic memory leak vulnerability in observable.rs
Add a 'shrink_amortized' fn to do an amortised O(1) capacity shrink of
the 'Observable::senders' private field whenever it is cleared of closed
channels, since there is no automatic shrinkage & realloc of std vectors
upon element removal. This fixes a hypothetical memory leak whereby a
linear number of observers could be attached to an 'ObservableHashMap'
key, then all but one are removed and added to another key, then all but
one are removed from that and added to the another, and so on, leading
to quadratic memory usage.