Use Unimock to mock nonempty wallet in integration tests
Make the wallet gRPC service give nontrivial 'RegisterConfidenceNtfn' &
'ListUnspent' responses by substituting a mock 'WalletService' impl for
some of the CLI tests, using the 'unimock' mocking crate. Feature gate
it as an optional instead of a dev dependency, so that mocks created for
'#[unimock]'-annotated traits are usable by the integration tests (that
is, across crates), as the lack of reflection in Rust means that extern
trait automocking is impossible without touching the production API. In
this way, the responses of the 'notify-confidence' & 'list-unspent' CLI
commands are better tested than by just using an empty wallet.

(Also run "cargo update" again and bump the minor version of Tokio, as
well as the patch versions of Tonic and a few transitive dependencies.)

(Unimock is currently being used instead of the more popular Mockall, as
the latter seems to generate a lot of macro-expanded code for its mocks,
and places more restrictions on the hierarchy of the mocked traits.)

TODO: Attempt to use Unimock's companion 'entrait' crate to better
 achieve IoC than with the current service architecture.