Add SubscribeTxConfirmationStatus RPC in line with Bisq2 changes
Sync 'rpc.proto' to the modified Bisq2 version, which adds a 'Musig'
service call to subscribe to deposit tx confirmation events, so that the
seller can pick those up, and not just the buyer from the (currently)
streamed result of 'PublishDepositTx'. This also allows the traders to
pick up the confirmation in case they go offline at the trade start, by
resubscribing upon app startup. Add a handler to 'server.rs' which
immediately streams a dummy tx confirmation event singleton, identical
to the current 'PublishDepositTx' dummy response.

Update the mock trade in 'TradeProtocolClient' to reflect this API
change, and add some comments to help clarify how the redaction of the
buyer's 'swapTxInputPartialSignature' is intended to work and the exact
trade phases each RPC call is made in.

(Also downgrade the Java release version in the POM from 22 to 21, to
bring it in line with Bisq2's JDK downgrade to 21.)

TODO: Consider using proto definitions and tx event subscription calls
 from 'walletrpc.proto' once the wallet service is better developed.