Make error enums non-exhaustive & add Clippy lint
Add the ('restriction'-group) 'exhaustive_enums' Clippy lint to
Cargo.toml, to catch any public enums which haven't been annotated as
'#[non_exhaustive]', as it may be reasonably expected that many of them
are error enums, which should not generally commit to a fixed set of
kinds. Accordingly, add the missing annotation to 'ProtocolErrorKind'
and 'WalletErrorKind'.