Factor out sync logic from WalletServiceImpl::connect
Add a private method, 'WalletServiceImpl::sync_from_rpc_emitter', to
deduplicate the code for the initial RPC sync and the polled syncing
that follows thereafter. Also change some of the log levels to TRACE and
ensure that the paired-up 'Wallet::apply_(evicted|unconfirmed)_txs'
method calls for the new (2.0) BDK API are carried out transactionally.