Append 'swapTxInputSighash' field to PartialSignaturesMessage proto
Add the swap-tx-input tap-sighash (that is, the raw message bytes being
signed for the swap tx input) to 'PartialSignaturesMessage' and the
corresponding 'ExchangedSigs' struct. This is to prepare for a small
security enhancement to the protocol, whereby the buyer learns only the
sighash of the swap tx input from the seller during the trade setup, and
the payout address of the swap tx is kept secret until its broadcast.
(This is OK because the buyer doesn't actually care what purported swap
tx is being signed, only that his adaptor will work for it.)

(This is a non-breaking change to the 'Musig' gRPC service, as the
sighash field is optional and set but not currently read by the server.)