Avoid passing secrets to the Java Musig client to hold
Add a '(buyer|seller)ReadyToRelease' flag to the request proto of the
'GetPartialSignatures' and 'SignSwapTx' Musig RPC methods, respectively.
Clearing these flags causes secrets to be redacted from the RPC service
responses, to be held securely by the Rust server & no longer expecting
the Bisq2 Java client to redact them from exchanged messages or hold on
to them until the buyer/seller is ready to release the escrow. Instead,
the client should call the respective methods a second time with the
flag enabled, once the trader has confirmed payment.

In addition to the logic to redact the respective buyer/seller secrets,
ignore all request proto fields besides the trade ID and the flag, for
any subsequent calls to 'GetPartialSignatures' or 'SignSwapTx', once the
partial signatures or signed swap tx have been determined, respectively.
This makes it a bit easier for the client to call the methods a second
time, which may be much later on during the trade.

Update 'TradeProtocolClient' to clarify how to use the modified API, as
this is a breaking change for the Bisq2 client.