Make buyer sign swap tx blindly with only peer-supplied sighash
Defer the buyer's partial signing of the swap tx to the next round, when
his deposit tx inputs are signed, in the 'SignDepositTx' Musig service
call. In this way, the buyer need not independently derive the swap-tx-
input sighash, and may instead use the sighash provided by the seller,
with the recently added 'PartialSignaturesMessage.swapTxInputSighash'
proto field. This fits in with the security enhancement of the previous
commit, to redact the buyer's swap-tx-input partial signature from the
server's 'GetPartialSignatures' RPC response, instead of expecting the
Java client to redact and hold it from the peer.

In this way, the swap tx is effectively signed blindly by the buyer, but
that is OK because the buyer only cares that his adaptor will function,
between the signature of anything spending the seller payout UTXO (but
not his own) and the full private key of his own payout UTXO. This must
be the case because at most one such tx was signed, namely the swap tx.

The main purpose of this change is to prevent nuisance-broadcast of the
swap tx by the buyer any time after a cooperative trade closure. That is
difficult to avoid if the buyer learns the unsigned swap tx, since the
adaptor is an affine congruence that necessarily works _both ways_
between the swap tx signature and the buyer payout private key.

NOTE: As a side effect of the above, the buyer must watch for any spend
 of the seller payout UTXO, when trying to detect the swap tx broadcast,
 instead of watching for a specific txid.