Upgrade Tonic/Prost to 0.14 & tidy/unify other dependencies
Update the versions of the 'prost' & 'tonic*' crates from 0.13 to 0.14.1
and 0.14.0 respectively. This is a breaking change which decouples Prost
from Tonic somewhat, requiring an extra 'tonic-prost' bridging crate and
'tonic-build' to be replaced with 'tonic-prost-build'. It also fixes the
'result_large_err' perf warning appearing with Rust 1.87, by boxing the
contents of 'tonic::Status', so remove the workaround to raise Clippy's
'large-error-threshold' config setting from 128 to 256.

Also re-add 'rand' as a direct dependency of the 'rpc' crate, instead of
importing it via 'musig2::secp256k1', since the 'secp256k1' crate is
just the _default_ crypto provider of 'musig2', which can also be made
to use the native-Rust 'k256' (though we're unlikely to ever want that).

Finally, add the shared dependencies 'anyhow' & 'rand' to the workspace
manifest, alongside BDK & MuSig2, to ensure that the 'rpc' & 'protocol'
packages are using the same versions. Also run "cargo update" and bump
Tokio from 1.46 to 1.47.0.