Add 'rust-2024-compatibility' lint & comments re. Rust 2024 migration
Enable the rustc lint to check for Rust 2021->2024 migration issues, but
hold the configured edition for 'rpc' at 2021 for now (unlike 'protocol'
which currently uses Rust 2024). This is because the (now-deprecated)
open source IDEA Rust plugin, currently maintained in a fork,

  https://github.com/l2dy/intellij-rust-unofficial-plugin

does not yet understand the "+ use<...>" syntax needed to make lifetime
captures of impl-traits precise, in order not to overcapture with the
modified rules of Rust 2024. This currently breaks 'rpc::observable', so
add explicit migration warning suppressions to that module.