Add taproot tweaks to all multisig outputs in rpc::protocol
Prevent a security issue, introduced by PR #75, whereby the seller can
trick his peer into signing an arbitrary tx spending the buyer's warning
tx output, by masquerading it as the (now blindly signed) swap tx. As
discussed in the PR, this can be mitigated by tweaking some or all of
the aggregated pubkeys, to ensure that the warning tx escrow outputs do
not share addresses/final-pubkeys with either of the deposit tx payouts.

To this end, give each of the four multisig output addresses a taproot
tweak, with the two deposit payouts tweaked with an empty unspendable
merkle root (preventing script spends), and the two warning outputs
tweaked with a merkle root giving a single script path each, namely the
respective time-locked claim paths. Thus, using a script spend in place
of a keyspend for the claim txs provides a convenient way to ensure that
all the multisig addresses in the protocol are distinct. Separate out
the TapTree computation logic to a new module, 'rpc::transaction'.

TODO: We can still provide (preferred, as cheaper and more private)
 keyspend paths for the claim txs, using prepared txs, with the script
 spends only as a backup in case of tx fee volatility, so add these to
 the mocked-up protocol and 'Musig' gRPC service API.