Add missing claim tx keyspend path to rpc::protocol
Add fields to 'NonceSharesMessage' and 'PartialSignaturesMessage' in
'rpc.proto' to construct a _prepared_ claim tx for each trader, making a
keyspend of his respective warning tx output. This requires passing an
additional claim payout address, input nonce shares & partial signatures
in Messages B, C & D during the trade setup, which would be a breaking
change to the 'Musig' gRPC API. To avoid the latter, simply log warnings
for missing/empty claim-tx-related proto fields in the 'Musig' requests,
for now, so that the final claim txs with their aggregated signatures
would be absent at the end of trade setup. (Making the prepared claim tx
construction optional is just a temporary measure to give the Bisq2 Java
client a chance to update the protocol.)

Thus, at the end of the trade setup, we get a pair of prepared claim txs
(with only one side possessing the full signature of his respective tx,
just as with the warning & redirect txs). But each trader also has a
backup script-spend claim path, with a dynamically constructed tx he can
use in place of his static claim tx (which makes a keyspend).

(All trade txs are mocked in the 'rpc' crate for now.)