Add tradeFeeReceiver & redirectionAmountMsat proto fields
Reuse the 'ReceiverAddressAndAmount' proto object to let the client
specify an optional trade fee recipient, as an extra request field of
'Musig.GetNonceShares'. This is the point in the trade setup where the
unsigned deposit & prepared txs are computed on the Rust side. Rename
the equivalent 'rpc::protocol::RedirectionReceiver' to just 'Receiver'
because of this reuse, and rename 'receivers' to 'redirectionReceivers'
in the proto.

Also provide 'Receiver::(total_)output_cost_msat' methods to calculate
the cost in millisatoshis of both the output amount and its mining fee
contribution. The latter is the reason for millisatoshi precision
(namely, to prevent the buildup of rounding errors over multiple such
receiver outputs, since we allow fractional tx feerates). Use this to
check that the absolute amounts in the client-supplied redirection
receiver list sum to the correct total, in the 'GetPartialSignatures'
handler. To aid the Java client, add an extra 'redirectionAmountMsat'
field to the response of the earlier 'GetNonceShares' call.

NOTE: For now, the receiver list validation just logs an error/warning
 upon failure, to give the Bisq2 client a chance to update.

Also update 'TradeProtocolClient' to use valid mock redirection amounts,
and correct a sats/vB to sats/kwu unit conversion error in the provided
mock feerates, making them 16x too high.