Add WarningTxBuilder to compute real sighashes
Build up real (unsigned) buyer & seller warning txs, in order to replace
some of the mock sighashes in 'rpc::protocol::sign_partial' with real
ones. To this end, add a 'WarningTxBuilder' struct to hold the params of
each warning tx, which starts off empty and consists of once-only-
initialised fields, enforced with 'Option::get_or_insert'. Add a pair of
buyer/seller-warning-tx builder fields to 'rpc::protocol::TradeModel',
following the pattern of the other 'TradeModel' fields, of having an
empty default and getting filled in as the trade progresses.

In order to avoid lots of boilerplate, define a 'make_getter_setter!'
macro to emit the needed builder accessors. This uses the 'paste' crate
to join tokens (specifically, prepending with 'set_' & 'Missing'), so
add it to the package manifest. Also bump the patch versions of 'tokio'
& 'clap' and run "cargo update" again.

Finally, add a unit test for 'WarningTxBuilder' to build an identical
(unsigned) warning tx to one pulled from a local Regtest blockchain,
after running the integration tests of the 'protocol' crate. NOTE: There
are currently a couple of minor divergences, needing temp code changes
to the 'protocol' crate to patch over:

 1) The integration-test warning txs opt out of RBF and locktime, so
    have no enforced delay; ours have a configurable relative locktime;
 2) The integration-test redirect txs have a relative locktime of one
    block at present; ours will have no locktime (but opt in to RBF).