Add RedirectTxBuilder to replace more mocks with real sighashes
Build up real unsigned buyer & seller redirect txs, in order to replace
some more of the mock sighashes in 'rpc::protocol::sign_partial' with
real ones. Similarly to 'WarningTxBuilder', add a pair of buyer/seller-
redirect-tx builder fields to 'rpc::protocol::TradeModel', following the
pattern of the other 'TradeModel' fields of having an empty default and
getting filled in as the trade progresses. Once all its params are set,
an unsigned redirect tx and associated input sighash can be computed.

Also add a unit test for the builder, to create an identical (unsigned)
redirect tx to the one pulled from a local Regtest blockchain (which was
included in the test data from an earlier commit), produced from a run
of the 'protocol' crate integration tests.