Prevent 'mismatched_lifetime_syntaxes' warning appearing with Rust 1.89
The new lint is triggered whenever a function has a return type with a
hidden lifetime that ties it to the lifetime of one of its arguments, as
that is misleading. Prevent the warning by adding an _explicitly_ elided
lifetime to the return type-expr, e.g. "Foo<'_, X>" instead of "Foo<X>".