Enforce that 'uint*' proto fields lie in signed range as well
Reject gRPC requests with unsigned numeric fields (amounts, fee rates,
etc.) that don't lie in the corresponding signed integer range as well,
such as uint64's that are too big to express as i64's. This is to help
prevent accidental wrapping when the API is consumed by Java clients, as
Java lacks unsigned primitives. (Also ensure that no _response_ contains
a numeric field that doesn't lie in the corresponding signed range.)

To this end, add a 'CheckInSignedRange' (extension) trait, implemented
only for u64 at present, with an extension method checking that 'self'
fits inside an i64.