Replace debug-formatting of request logs with JSON
Make the logging of requests (for the 'Musig' & 'Wallet' gRPC services)
a little more tidy and compact, by formatting the request data as JSON
using Serde, in place of Rust's debug formatting. This is in a similar
vein to the change to JSON formatting of the CLI output made in the much
earlier commit 0c16148.

To this end, add the relevant proto types and fields-to-annotate to the
Prost configuration in 'build.rs', alongside the existing ones for the
'walletrpc' responses used by the CLI. Also add helpers to serialise
byte-array fields as base64 (which is closer to the canonical encoding
of protobuf as JSON) instead of just hex, and use base64 encoding for
the key share, nonce, partial signature and PSBT fields of the 'Musig'
request protos. (Keep hex encoding for other bitcoin-consensus-related
fields though, such as txids & scriptPubKeys, as it's less confusing.)

(Also use a private 'LazyJson' helper struct to make sure that there's
no performance hit to serialise objects as JSON if they aren't visible
at the given log level.)