Pretty print debug-formatted structured log fields
Configure 'tracing-subscriber' to use alternative formatting (that is,
"{:#?}" in place of "{:?}") for all debug-formatted log fields, such as
the ('LazyJson'-wrapped) gRPC request data. This makes the logs a bit
more verbose (at least at the default DEBUG level of the '::rpc' module
root), but makes the logged requests a lot easier to read, which should
aid testing and development.