Add streamed & un-streamed gRPC response logging at TRACE level
In addition to the RPC request messages, which are logged (as JSON) at
DEBUG level, log the streamed & un-streamed response messages (as JSON)
at the TRACE level (which is invisible by default). This requires every
response proto to implement 'Serialize', so add the extra types & fields
to the Prost configuration in 'build.rs'.

It also requires the associated response Stream types of the 'MusigImpl'
and 'WalletImpl' service trait impls to implement 'Serialize'. To this
end, add a "TracedResultStream<T>: Serialize" struct to wrap the present
associated types of the form "BoxStream<'static, Result<T>>", containing
a 'span' field to reenter the current Tracing span (at the point of
construction) each time the stream is polled, for better logging. Add a
'box_traced' extension method to box a Stream as a 'TracedResultStream',
with each streamed response item (proto) logged at TRACE level. (The
wrapper stream itself serialises to an empty "{}" JSON response, for
convenience to the code.)

Finally, make the 'LazyJson' helper newtype implement Display in place
of Debug, as it turns out that switching to alternative field-formatting
affects both Display (%field) & Debug (?field) output, in spite of what
the 'tracing-subscriber' documentation misleadingly implies, and custom
_Display_ output is probably more appropriate.