Set missed fee receiver & deposit PSBT fields in TradeProtocolClient
Make sure to initialise 'NonceSharesRequest.tradeFeeReceiver' with a
meaningful address+amount in the mock-trade-protocol Java tests, instead
of leaving it null. Also set a missed 'peersDepositPsbt' request field,
(renamed from 'depositPsbt' in the proto) when calling PublishDepositTx.
Its absence was not detected by the Rust server, as the handler is just
an unimplemented stub at present, so accordingly add server-side checks
that it is a non-null, valid PSBT. (The Bisq2 client already correctly
forwards the peer's deposit PSBT, unlike our reference Java code.)

Finally, remove 'SubscribeTxConfirmationStatusRequest.depositPsbt' from
'rpc.proto', since the peer's signatures on the tx aren't needed by the
server just to attach a listener for it -- only the trade ID.