Add signatures to prepared-tx-builders & pass real Swap Tx to client
Extend '(Forwarding|Redirect|Warning)TxBuilder' with getters & setters
for (keyspend) signatures on each input, a field + getter for the final
signed tx & a 'compute_signed_tx(&mut self)' method to populate it. The
task of actually computing (or validating) the signatures is not deemed
to be the builders' responsibility (and it would be a little wasteful to
check them twice, since the 'musig2' library already ensures validity).

Also add a 'WithWitnesses' (extension) trait for 'bitcoin::Transaction',
with a 'with_key_spend_witness' fluent method to initialise the witness
at a given input index with a given keyspend signature. It also adds a
'find_key_spend_signature' method, to scan all the inputs for a keyspend
signature against a given matching outpoint. This is needed by the buyer
to recover the Swap Tx signature, and hence his full payout private key,
since the structure of the Swap Tx is arbitrary as far as he's concerned
-- all he knows is that it makes a keyspend of the seller's payout UTXO.

Using the new builder functionality, together with the 'WithWitnesses'
trait, replace the Swap Tx's final signature in the mocked-up Musig gRPC
API with a real signed Swap Tx, instead of having the former masquerade
as the latter for test/development of the protocol crypto.