Enforce that all Deposit Tx p2tr inputs are keyspend-only
Use the optional 'tap_internal_key' field of each half-PSBT input to
ensure that it has no script-spend path, by recomputing the tweaked
pubkey (with empty merkle root) to check for a match against its SPK.
This will help to prevent unexpected severe underpayment of the Deposit
Tx fee, as the satisfaction weight of a script-spending input is hard to
predict. To this end, ensure that the 'tap_internal_key' field of each
mock PSBT input is populated, and add an 'Option<XOnlyPublicKey>' field
to 'TxOutput' to hold an internal key for the given coin, if known.

Also add a check that each half-deposit PSBT input has sequence number
0xFFFFFFFD (enable RBF, no-relative-locktime) & empty witness+ScriptSig
(with the latter required for any well-formed PSBT -- this is important
for security to prevent downstream prepared txs from being created with
invalid Deposit Tx outpoints).