Add missing placeholder/receiver output checks for deposit PSBT halves
Make sure that the peer's OP_RETURN placeholder output & possible trade
fee receiver outputs (in the case of the seller) are well-formed and set
to the correct amounts & receiver address(es). This should prevent theft
of Deposit Tx network fees and ensure that the seller actually pays the
agreed BM the required trade fee (if specified).