Add NetworkParams trait to supply prepared-tx locktimes
Provide an 'rpc::transaction::NetworkParams' extension trait to supply
network-specific tx params to the protocol, extending 'bitcoin::Network'
with the methods '(warning|redirect|claim)_lock_time(&self)'. In this
way, all the remaining Regtest-specific code can be removed from the
'rpc::protocol' module. Add new 'MAINNET_*' constants to 'transaction'
for the default prepared-tx locktimes on networks besides Regtest.
Currently, this is a flat 10-day (1440-block) Warning Tx delay and a
5-day (720-block) Claim Tx delay, which may need further consideration.

Also add a (relative-)locktime setter to 'RedirectTxBuilder' for extra
flexibility, although this is currently set to 'LockTime::ZERO' for all
networks, which disables it. (That is, the input sequence number is set
to 0xFFFFFFFD, as before).