Move tx builder logic from 'rpc' to 'protocol' crate
Move the 'transaction', 'psbt' & 'swap' modules to the 'protocol' crate,
making the former two public (as necessary, for access by the unmoved
'rpc::protocol' module). As this caused a lot of 'missing_errors_doc'
(pedantic-group) Clippy warnings to appear, disable the lint globally,
since it's probably not very useful anyway for not-yet-mature code.

Also move/copy some external dependencies to the 'protocol' package, to
get it to build, and tighten the access of some of the definitions in
'transaction' & 'psbt' that can be (crate-)private since they are only
shared between those two modules.