Implement TradeWallet trait for bdk_wallet::Wallet
Provide a BDK wallet impl for 'protocol::psbt::TradeWallet', so that the
protocol tx builders can construct a Deposit Tx which spends funds from
a real wallet, as well as prepared txs which return funds to it.

Also provide a unit test for the 'TradeWallet::create_half_deposit_psbt'
and 'TradeWallet::sign_selected_inputs' method impls for 'Wallet', which
are nontrivial. The test uses a fake wallet provided by the feature-
gated BDK wallet test utilities, funded with 50_000 sats, so enable this
as a dev-dependency (only) in 'protocol/Cargo.toml'.