Use WarningTxBuilder in protocol_musig_adaptor::WarningTx
Replace the existing Warning Tx building logic with calls to the tx
builders, updating the 'build', 'build_partial_sig' & 'aggregate_sigs'
methods of 'WarningTx' appropriately. In particular, use the previously
untested sighash calculations of 'WarningTxBuilder' when generating
partial signatures, instead of using the sighash calculation bundled in
each call to 'TMuSig2::generate_partial_sig'.

Also hold the generated Warning Tx in the builder instead of directly in
the 'WarningTx' struct, replacing its 'get_tx' method with 'signed_tx' &
'unsigned_tx' getters which delegate to those of the wrapped builder.

As with the previous commit, this slightly changes the form of the
resultant Warning Tx, making both its inputs RBF opt-in, with sequence
0xFFFFFFFD instead of 0xFFFFFFFF, and using a precise feerate (currently
hardcoded to 10 sat/vB).