Move protocol::transaction::Receiver to its own module
Add the pub module 'protocol::receiver' to hold the 'Receiver' struct of
trade-fee/redirection receivers, to keep the size of the 'transaction'
module down. This is in preparation for placing more receiver-related
logic there, including a possible port of algorithm used in Bisq2 to
calculate the absolute receiver amounts from the specified BM shares.