Group imports in rpc & protocol crates by 'StdExternalCrate'
Rearrange the imports in the 'rpc' & 'protocol' source files slightly,
grouping them in accordance with the 'rustfmt.toml' config setting:

  group_imports = "StdExternalCrate",

for consistency across the project.

Exclude the (style_)edition = "2021" settings from the config, however,
as that overrides the Rust 2021/2024 edition setting picked up from the
Cargo manifest, which changes the import orders for 'protocol' slightly.