Fix incorrect weight estimate of P2WPKH inputs
Set the return value of 'TxOutput::estimated_input_weight' to 271 wu
instead of 272 wu for P2WPKH inputs, as a standard low-R bitcoin ECSDA
signature is (up to) 71 bytes including the sighash type, not 72 bytes.

Also remove the 'dbg!(sighash)' lines and corresponding TODOs from the
tx builder unit tests in 'protocol::transaction', as the correctness of
the builder sighashes is already being tested indirectly by the protocol
integration tests in 'lib.rs'.