Add semaphore to make protocol integration tests run in series
Use 'tokio::sync::Semaphore' to enforce serial execution of the (Nigiri-
dependent) integration tests in 'protocol/src/lib.rs', so that
'--test-threads=1' no longer needs to be passed as a command-line arg to
"cargo test". The semaphore from Tokio was chosen because it is fair
(that is, it has FIFO permit acquisition), reducing the tendency to get
warnings that a given test has been running for more than 60 seconds
(though they presently seem to occur regardless, as the tests are slow).

Remove the '#[test]' annotation from 'protocol::nigiri::check_start' and
run it as a delegated async test in 'lib.rs' using '#[tokio::test]',
which creates an async context to await the semaphore permit & rewrites
the test as a synchronous fn callable by all the other tests in the
module. In this way, the permit is acquired each time 'check_start' is
called and released at the end of each test.

Also add 'tokio' to the dev-dependencies of 'protocol' and the workspace
manifest (as it is also a dependency of 'rpc').