Move protocol integration tests into 'protocol/tests' dir
Move the end-to-end protocol tests out of 'lib.rs' and into their own
file under 'tests', as this is the standard place for integration tests.

Since each such file is compiled as a separate crate by "cargo test",
widen the access of some of the fields & fns in 'protocol_musig_adaptor'
as appropriate, from crate-public to public (namely, some of the fields
of 'BMPProtocol' & 'AggKey', and the 'broadcast' method of each prepared
tx wrapper struct).

(Also update the README, as passing "--lib tests" instead of '--lib' &
'--tests' as separate command-line options to "cargo test" doesn't look
correct and seems to exclude the integration tests from running.)