Split 'KeyPair<PrvKey: ValStorage>' into 2 monomorphic types
Simplify 'rpc::protocol::KeyPair' by creating a separate similar struct,
'OptKeyPair', with an optional 'prv_key' field, instead of making the
type polymorphic over the optionality of 'prv_key' using 'rpc::storage'.
This introduces a small amount of duplication but is probably clearer.

Also remove 'rpc::storage::{ValStorage, ByOptVal}', as it seems that
abstraction over the optionality of a struct field is not very helpful,
only being used in one place. (If a better use case arises, it could be
put back in.)