Move (Key|Sig)Ctx from rpc::protocol to new module in protocol crate
Attempt to split off the multisig logic in 'rpc::protocol' into a new
'protocol::multisig' module, which includes the structs '(Key|Sig)Ctx',
'(Key|OptKey|Nonce)Pair' and their associated impls. This is with the
aim of addressing any differences with the independent multisig logic in
'protocol_musig_adaptor' and making both use the same code.

Don't make any other code changes for now, besides widening visibility
of structs, fields & fns as needed, splitting off a separate error enum
for the new module and replacing "result" with "Ok(result?)" in a couple
of cases to get the code to compile.