Make key share order data-dependent, to remove 'KeyCtx::am_buyer' field
Simplify the 'multisig::KeyCtx' API, by switching from a role-dependent
(buyer-then-seller) key share aggregation order to a data-dependent
(least-pubkey-then-greatest-pubkey) order, in agreement with the method
used in 'protocol_musig_adaptor'. (This differs from the approach taken
in Bisq1, where multisig keys are ordered by role in the escrow redeem
script, but that doesn't seem to offer any obvious benefits for MuSig2.)

This allows the now-unused 'am_buyer' field of 'KeyCtx' to be removed.