Make KeyCtx fields private and merge KeyPair & OptKeyPair
Simplify and improve the safety of the 'KeyCtx/KeyPair/OptKeyPair' API,
by making all their fields private (with accessors preserving all their
invariants) and merging the latter two into a single 'KeyPair' struct.
Also ensure that the initialisation of all the nested 'KeyCtx' fields is
once-only (with any subsequent mutation attempt a NO-OP).

As part of this, add a 'MultisigErrorKind::MissingPrvKey' error variant,
to indicate that a given keypair share / aggregated keypair is present
but the private (scalar) component of it is unknown/unset.