Improve safety of SigCtx API & remove unnecessary zero nonce check
Remove the check that the aggregated nonce isn't zero (which can only
occur when a peer is misbehaving), along with the explaining comment.
The latter was in error, since it failed to account for the fact that a
peer cannot make the aggregated nonce zero and still produce a valid
partial signature. (Also, as explained in BIP 327, the nonce defaulting
added to the MuSig2 scheme to prevent early failure in that case doesn't
weaken its security, as it's reducible to the unmodified scheme that
cannot tolerate a zero nonce.)

Also add a missing check to 'SigCtx::compute_taproot_signature' that the
supplied adaptor secret matches the adaptor point, which is a necessary
and sufficient condition for the final signature to be valid, given a
valid aggregated adaptor signature.

Disallow the adaptor point to be set after the local nonce is (randomly)
initialised, instead of merely preventing modification after partially
signing a message. This is as a precaution, since the MuSig2 + adaptors
scheme doesn't appear to have a formal specification or security proof.
But initializing the nonce only after the adaptor point is decided would
seem to allow the adaption to be modelled as a fake third party adding
his partial signature, with a dummy private key of zero and a secret
nonce of the form (s, 0) where 's' is the adaptor secret. In this way,
provable security could likely be obtained.

Finally, replace the 'SigCtx' private methods of the form 'get_*', as
that is not an idiomatic Rust naming convention for getters.